backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
boto3==1.29.7
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.1
requests==2.31.0
aiofiles==23.2.1
httpx==0.25.1
onnxruntime==1.16.0
torch==2.9.1
pinecone-client==3.1.0
apscheduler==3.10.4

# Face Recognition and Emotion Detection
face-recognition==1.3.0
dlib==19.24.2
deepface==0.0.79
tensorflow==2.15.0
scikit-learn==1.3.2